--- Source/SimpleTextTemplate.Renderer.Core/SimpleTextTemplate.Renderer.Core.csproj
@@ -1,5 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-
-
-</Project>
