--- Source/SimpleTextTemplate.Renderer/SimpleTextTemplate.Renderer.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <IsPackable>true</IsPackable>
-    <Description>SimpleTextTemplateのレンダリング部分です。</Description>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageTags>templating</PackageTags>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SimpleTextTemplate.Abstractions\SimpleTextTemplate.Abstractions.csproj" />
-    <ProjectReference Include="..\SimpleTextTemplate.Renderer.Core\SimpleTextTemplate.Renderer.Core.csproj" />
-    <ProjectReference Include="..\SimpleTextTemplate\SimpleTextTemplate.csproj" />
-  </ItemGroup>
-
-</Project>
