+++ Tests/SimpleTextTemplate.Writer.Tests/SimpleTextTemplate.Writer.Tests.csproj
