+++ Tests/SimpleTextTemplate.Writer.Tests/TemplateWriterWriteConstantLiteralTest.cs
@@ -3,7 +3,7 @@ using System.Text;
 using FluentAssertions;
 using Xunit;
 
-namespace SimpleTextTemplate.Renderer.Core.Tests;
+namespace SimpleTextTemplate.Writer.Tests;
 
 public sealed class TemplateWriterWriteConstantLiteralTest
 {
