+++ Tests/SimpleTextTemplate.Writer.Tests/TemplateWriterWriteEnumTest.cs
@@ -2,7 +2,7 @@
 using FluentAssertions;
 using Xunit;
 
-namespace SimpleTextTemplate.Renderer.Core.Tests;
+namespace SimpleTextTemplate.Writer.Tests;
 
 public sealed class TemplateWriterWriteEnumTest
 {
