+++ Tests/SimpleTextTemplate.Writer.Tests/TemplateWriterWriteLiteralTest.cs
@@ -3,47 +3,49 @@ using System.Text;
 using FluentAssertions;
 using Xunit;
 
-namespace SimpleTextTemplate.Renderer.Core.Tests;
+namespace SimpleTextTemplate.Writer.Tests;
 
-public sealed class TemplateWriterWriteStringTest
+public sealed class TemplateWriterWriteLiteralTest
 {
     [Theory]
-    [InlineData(null)]
     [InlineData("")]
     [InlineData("a")]
     [InlineData("abc01234567890")]
-    public void 文字列_バッファーライターに書き込み(string? value)
+    public void 文字列_バッファーライターに書き込み(string value)
     {
+        var utf8Value = Encoding.UTF8.GetBytes(value);
         var bufferWriter = new ArrayBufferWriter<byte>();
 
         var writer = TemplateWriter.Create(bufferWriter);
-        writer.WriteString(value);
+        writer.WriteLiteral(utf8Value);
         writer.Flush();
 
         bufferWriter.WrittenSpan.ToArray()
             .Should()
-            .Equal(Encoding.UTF8.GetBytes(value ?? string.Empty));
+            .Equal(utf8Value);
     }
 
     [Fact]
     public void 長い文字列_バッファーライターに書き込み()
     {
         var value = new string('a', 1024);
+        var utf8Value = Encoding.UTF8.GetBytes(value);
         var bufferWriter = new ArrayBufferWriter<byte>();
 
         var writer = TemplateWriter.Create(bufferWriter);
-        writer.WriteString(value);
+        writer.WriteLiteral(utf8Value);
         writer.Flush();
 
         bufferWriter.WrittenSpan.ToArray()
             .Should()
-            .Equal(Encoding.UTF8.GetBytes(value));
+            .Equal(utf8Value);
     }
 
     [Fact]
     public void 文字列を複数回追加_バッファーライターに書き込み()
     {
         var value = new string('a', 30);
+        var utf8Value = Encoding.UTF8.GetBytes(value);
         var bufferWriter = new ArrayBufferWriter<byte>();
         var count = 0;
 
@@ -51,7 +53,7 @@ public sealed class TemplateWriterWriteStringTest
 
         for (; count < 10; count++)
         {
-            writer.WriteString(value);
+            writer.WriteLiteral(utf8Value);
         }
 
         writer.Flush();
