+++ Tests/SimpleTextTemplate.Writer.Tests/TemplateWriterWriteStringTest.cs
@@ -3,49 +3,47 @@ using System.Text;
 using FluentAssertions;
 using Xunit;
 
-namespace SimpleTextTemplate.Renderer.Core.Tests;
+namespace SimpleTextTemplate.Writer.Tests;
 
-public sealed class TemplateWriterWriteLiteralTest
+public sealed class TemplateWriterWriteStringTest
 {
     [Theory]
+    [InlineData(null)]
     [InlineData("")]
     [InlineData("a")]
     [InlineData("abc01234567890")]
-    public void 文字列_バッファーライターに書き込み(string value)
+    public void 文字列_バッファーライターに書き込み(string? value)
     {
-        var utf8Value = Encoding.UTF8.GetBytes(value);
         var bufferWriter = new ArrayBufferWriter<byte>();
 
         var writer = TemplateWriter.Create(bufferWriter);
-        writer.WriteLiteral(utf8Value);
+        writer.WriteString(value);
         writer.Flush();
 
         bufferWriter.WrittenSpan.ToArray()
             .Should()
-            .Equal(utf8Value);
+            .Equal(Encoding.UTF8.GetBytes(value ?? string.Empty));
     }
 
     [Fact]
     public void 長い文字列_バッファーライターに書き込み()
     {
         var value = new string('a', 1024);
-        var utf8Value = Encoding.UTF8.GetBytes(value);
         var bufferWriter = new ArrayBufferWriter<byte>();
 
         var writer = TemplateWriter.Create(bufferWriter);
-        writer.WriteLiteral(utf8Value);
+        writer.WriteString(value);
         writer.Flush();
 
         bufferWriter.WrittenSpan.ToArray()
             .Should()
-            .Equal(utf8Value);
+            .Equal(Encoding.UTF8.GetBytes(value));
     }
 
     [Fact]
     public void 文字列を複数回追加_バッファーライターに書き込み()
     {
         var value = new string('a', 30);
-        var utf8Value = Encoding.UTF8.GetBytes(value);
         var bufferWriter = new ArrayBufferWriter<byte>();
         var count = 0;
 
@@ -53,7 +51,7 @@ public sealed class TemplateWriterWriteLiteralTest
 
         for (; count < 10; count++)
         {
-            writer.WriteLiteral(utf8Value);
+            writer.WriteString(value);
         }
 
         writer.Flush();
